backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
pydantic-settings>=2.2.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.9
httpx>=0.26.0
pytest>=8.0.0
web3>=6.15.0
PyMuPDF>=1.24.0
pdfminer.six>=20231228
